fix(sync): wire dirty-since guard so pull loop cannot overwrite local edits
set_dirty_since() was implemented but never called, meaning the pull-loop
guard (skip if dirty_since >= remote_ts) could never trigger. Any pull cycle
could silently overwrite an in-progress local edit.

Fixes:
- Call set_dirty_since(ino, Some(now_ms)) in DbFile::write() and in
  UnisonFs::enqueue_write() so every local write arms the guard atomically
  with queuing the push.
- Clear dirty_since on a successful push (push loop) and on a clean mirror
  (pull loop after applying a remote update) so the flag doesn't stay set
  permanently.
- Wire set_mirrored_state() into both loops: pull marks inodes "ok" after
  mirroring; push marks them "pushed" after upload.
- Wire sync_meta_get/set() into pull and push loops to persist last-pull-at
  and last-push-at timestamps for diagnostics.
- Wire delete_remote_path() into enqueue_delete() so the remote-path mapping
  is cleaned up when a file is unlinked.

All five previously-dead db.rs methods now have real call sites; the
dead_code warning is gone without any #[allow(dead_code)] suppression.

Also fix five pre-existing clippy lints in file.rs and nfs.rs (unnecessary
casts, or_insert_with → or_default, manual saturating_sub).

+++ crates/unisonfs-core/src/mount/nfs.rs
@@ -399,7 +399,7 @@ fn vfs_attr_to_nfs(attr: &crate::vfs::FileAttr) -> nfsserve::nfs::fattr3 {
 
     fattr3 {
         ftype,
-        mode: (attr.mode & 0o7777) as u32,
+        mode: attr.mode & 0o7777,
         nlink: attr.nlink,
         uid: attr.uid,
         gid: attr.gid,
